requirements.txt
python.dotenv==1.1.0
streamlit==1.45.0
openai==1.76.2
langsmith=0.3.40